Fix: Remove hidden row from texture
That extra row was actually used to create empty lines, but we could
just as well have used the space character (stretched to be a whole line
in length). Which is what we are now doing.

That change open up the posibility to have scrolling and screen clearing
take `is_inversed` into account.