Fix: Refreshing content texture sometimes fails
This has only once been witnessed, but the way it looked, it could not
have been other than the result of the texture not been cleared before
refreshing (i.e. there were some artifacts on the border of the texture.
Usually nothing is ever drawn on that border, meaning that if for some
reason it gets "damaged", then the only way to fix that would be to
clear the whole texture before redrawing on it).